Implement default traits on DocumentStore
