+++ p2panda-rs/src/test_utils/db/mod.rs
@@ -11,6 +11,4 @@ mod types;
 pub mod validation;
 
 pub use provider::MemoryStore;
-pub use types::{
-    EntryArgsResponse, PublishEntryResponse, PublishedOperation, StorageEntry,
-};
+pub use types::{EntryArgsResponse, PublishEntryResponse, PublishedOperation, StorageEntry};
