+++ p2panda-rs/src/test_utils/db/stores/mod.rs
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-mod document;
 mod entry;
 mod log;
 pub mod operation;
